Moving solve methods to problem class
