+++ examples/slot_3d/analysis.py
@@ -32,11 +32,6 @@ FEASolver = pyTACS(bdfFile, comm, options=structOptions)
 # Don't need a elemCallBack since property info exists in bdf
 FEASolver.createTACSAssembler()
 
-# Add Functions
-FEASolver.addFunction('mass', functions.StructuralMass)
-FEASolver.addFunction('ks_vmfailure', functions.KSFailure, safetyFactor=1.5,
-                      ksWeight=100.0)
-
 # ==============================================================================
 # Setup static problem
 # ==============================================================================
@@ -45,22 +40,28 @@ evalFuncs = ['mass', 'ks_vmfailure']
 # Read in forces from BDF and create tacs static problems
 SPs = FEASolver.createTACSProbsFromBDF()
 
+# Set up eval functions
+for spID in SPs:
+    SPs[spID].addFunction('mass', functions.StructuralMass)
+    SPs[spID].addFunction('ks_vmfailure', functions.KSFailure, safetyFactor=1.5,
+                          ksWeight=100.0)
+
 # Solve state
 for spID in SPs:
-    FEASolver(SPs[spID])
+    SPs[spID].solve()
 
 # Evaluate functions
 funcs = {}
 for spID in SPs:
-    FEASolver.evalFunctions(SPs[spID], funcs, evalFuncs=evalFuncs)
+    SPs[spID].evalFunctions(funcs, evalFuncs=evalFuncs)
 
 if comm.rank == 0:
     pprint(funcs)
 
 funcsSens = {}
 for spID in SPs:
-    FEASolver.evalFunctionsSens(SPs[spID], funcsSens, evalFuncs=evalFuncs)
+    SPs[spID].evalFunctionsSens(funcsSens, evalFuncs=evalFuncs)
 if comm.rank == 0:
     pprint(funcsSens)
 
-FEASolver.writeSolution(outputDir=os.path.dirname(__file__))
+SPs[spID].writeSolution(outputDir=os.path.dirname(__file__))
