+++ tacs/problems/__init__.py
@@ -1,5 +1,7 @@
 from .static import StaticProblem
+from .transient import TransientProblem
 
 __all__ = [
     "StaticProblem",
+    "TransientProblem",
 ]
