+++ tacs/pytacs.py
@@ -184,15 +184,10 @@ class pyTACS(BaseUI):
         self.globalDVs = {}
         self.compIDBounds = {}
         self.addedCompIDs = set()
-        self.varName = 'struct'
-        self.coordName = 'Xpts'
 
-        self.curSP = None
         self.doDamp = False
-        self._factorOnNext = True
         self._PCfactorOnNext = False
-        # List of functions
-        self.functionList = OrderedDict()
+
         self.adjointList = OrderedDict()
         self.dIduList = OrderedDict()
         self.dvSensList = OrderedDict()
@@ -203,13 +198,6 @@ class pyTACS(BaseUI):
 
         # Variables per node for model
         self.varsPerNode = None
-
-        # Norms
-        self.initNorm = 0.0
-        self.startNorm = 0.0
-        self.finalNorm = 0.0
-        # Flag for mat/vector creation
-        self._variablesCreated = False
 
         # TACS assembler object
         self.assembler = None
@@ -460,72 +448,6 @@ class pyTACS(BaseUI):
 
         return compIDs
 
-    def addFunction(self, funcName, funcHandle, include=None, exclude=None,
-                    includeBound=None, compIDs=None, **kwargs):
-        """
-        Generic function to add a function for TACS. It is intended to
-        be reasonably generic since the user supplies the actual
-        function handle to use. The following functions can be used:
-        KSFailure, KSBuckling, MaxBuckling, AverageKSFailure,
-        MaxFailure, AverageMaxFailure, AverageKSBuckling,
-        StructuralMass, Compliance, AggregateDisplacement.
-
-        Parameters
-        ----------
-        funcName : str
-            The user-supplied name for the function. This will
-            typically be a string that is meanful to the user
-        funcHandle : tacs.functions
-
-            The fucntion handle to use for creation. This must come
-            from the functions module in tacs.
-
-        include : varries
-            Argument passed to selctCompIDs. See this function for
-            more information
-
-        exclude : varries
-            Argument passed to selctCompIDs. See this function for
-            more information
-
-        compIDs: list
-            List of compIDs to select. Alternative to selectCompIDs
-            arguments.
-        """
-
-        # First we will get the required domain, but only if both
-        # include is None and exclude is None. If so, just use the
-        # entire domain:
-
-        # Note nGroup is one since we only want exactly one domain
-        if compIDs is None:
-            compIDs = self.selectCompIDs(include, exclude, includeBound,
-                                         nGroup=1)[0]
-
-        # Flatten and get element numbers on each proc corresponding to specified compIDs
-        compIDs = self._flatten(compIDs)
-        elemIDs = self.meshLoader.getLocalElementIDsForComps(compIDs)
-
-        # We try to setup the function, if it fails it may not be implimented:
-        try:
-            # pass assembler an function-specific kwargs straight to tacs function
-            self.functionList[funcName] = funcHandle(self.assembler, **kwargs)
-        except:
-            self.TACSWarning("Function type %s is not currently supported "
-                        "in pyTACS. Skipping function." % funcHandle)
-            return
-
-        # Finally set the domain information
-        self.functionList[funcName].setDomain(elemIDs)
-
-        # Create additional tacs BVecs to hold adjoint and sens info
-        self.adjointList[funcName] = self.assembler.createVec()
-        self.dIduList[funcName] = self.assembler.createVec()
-        self.dvSensList[funcName] = self.assembler.createDesignVec()
-        self.xptSensList[funcName] = self.assembler.createNodeVec()
-
-        return compIDs
-
     def getCompNames(self, compIDs):
         """
         Return a list of component descriptions for the given component
@@ -547,61 +469,6 @@ class pyTACS(BaseUI):
             compDescripts.append(self.compDescripts[compIDs[i]])
 
         return compDescripts
-
-    def getFunctionKeys(self):
-        """Return a list of the current function key names"""
-        return list(self.functionList.keys())
-
-    def setStructProblem(self, structProblem):
-        """Set the structProblem. This function can be called by the
-        user but typically will be called automatically by functions
-        that accept a structProblem object.
-
-        Parameters
-        ----------
-        structProblem : instance of pyStruct_problem
-            Description of the sturctural problem to solve
-            """
-
-        if structProblem is self.curSP:
-
-            return
-
-        if self.comm.rank == 0:
-            print('+' + '-' * 70 + '+')
-            print('|  Switching to Struct Problem: %-39s|' % structProblem.name)
-            print('+' + '-' * 70 + '+')
-
-        try:
-            structProblem.tacsData
-        except AttributeError:
-            structProblem.tacsData = TACSLoadCase()
-            structProblem.tacsData.F = self.assembler.createVec()
-            structProblem.tacsData.u = self.assembler.createVec()
-            structProblem.tacsData.auxElems = tacs.TACS.AuxElements()
-
-        # We are now ready to associate self.curSP with the supplied SP
-        self.curSP = structProblem
-        self.curSP.adjointRHS = None
-        # Force and displacement vectors for problem
-        self.F = self.curSP.tacsData.F
-        self.u = self.curSP.tacsData.u
-        # Set auxiliary elements for adding tractions/pressures
-        self.auxElems = self.curSP.tacsData.auxElems
-        self.assembler.setAuxElements(self.auxElems)
-
-        # Create numpy array representation for easier access to vector values
-        vpn = self.varsPerNode
-        self.F_array = self.F.getArray()
-        self.u_array = self.u.getArray()
-        self.F_array = self.F_array.reshape(len(self.F_array) // vpn, vpn)
-        self.u_array = self.u_array.reshape(len(self.u_array) // vpn, vpn)
-
-        # Set current state variables in assembler
-        self.assembler.setVariables(self.u)
-
-        # Reset the Aitken acceleration for multidisciplinary analyses
-        self.doDamp = False
 
     def createTACSAssembler(self, elemCallBack=None):
         """
@@ -653,11 +520,15 @@ class pyTACS(BaseUI):
 
         self.assembler = self.meshLoader.createTACSAssembler(self.varsPerNode)
 
-        self._createVariables()
         self._createOutputViewer()
 
+        # Initial design variable vec if necessary
+        self.dvs0 = self.assembler.createDesignVec()
+        self.assembler.getDesignVars(self.dvs0)
+
         # Initial set of nodes for geometry manipulation if necessary
-        self.coords0 = self.getCoordinates()
+        self.coords0 = self.assembler.createNodeVec()
+        self.assembler.getNodes(self.coords0)
 
     def _elemCallBackFromBDF(self):
         """
@@ -1385,7 +1256,7 @@ class pyTACS(BaseUI):
                 name = subCase.params['SUBTITLE'][0]
             else:
                 name = 'load_set_%.3d' % (subCase.id)
-            sp = tacs.problems.static.StaticProblem(name=name)
+            staticProb = self.createStaticProblem(name)
 
             if 'LOAD' in subCase.params:
                 loadsID = subCase.params['LOAD'][0]
@@ -1402,25 +1273,25 @@ class pyTACS(BaseUI):
                             loadArray[:3] += scale * loadInfo.scaled_vector
                         elif loadInfo.type == 'MOMENT' and vpn >= 6:
                             loadArray[3:6] += scale * loadInfo.scaled_vector
-                        self.addLoadToNodes(sp, nodeID, loadArray, nastranOrdering=True)
+                        staticProb.addLoadToNodes(nodeID, loadArray, nastranOrdering=True)
 
                     # Add any pressure loads
                     # Pressure load card specific to shell elements
                     elif loadInfo.type == 'PLOAD2':
                         elemIDs = loadInfo.eids
                         pressure = scale * loadInfo.pressure
-                        self.addPressureToElements(sp, elemIDs, pressure, nastranOrdering=True)
+                        staticProb.addPressureToElements(elemIDs, pressure, nastranOrdering=True)
 
                     # Alternate more general pressure load type
                     elif loadInfo.type == 'PLOAD4':
-                        self._addPressureFromPLOAD4(sp, loadInfo, scale)
+                        self._addPressureFromPLOAD4(staticProb, loadInfo, scale)
 
                     else:
                         self.TACSWarning("Unsupported load type "
                                     " '%s' specified for load set number %d, skipping load" %(loadInfo.type, loadInfo.sid))
 
             # append to list of structural problems
-            structProblems[subCase.id] = sp
+            structProblems[subCase.id] = staticProb
 
         return structProblems
 
@@ -1476,592 +1347,12 @@ class pyTACS(BaseUI):
 
             # Figure out whether or not this is a traction based on if a vector is defined
             if np.linalg.norm(loadInfo.nvector) == 0.0:
-                self.addPressureToElements(staticProb, elemID, pressure, faceIndex,
+                staticProb.addPressureToElements(elemID, pressure, faceIndex,
                                            nastranOrdering=True)
             else:
                 trac = pressure * loadInfo.nvector
-                self.addTractionToElements(staticProb, elemID, trac, faceIndex,
+                staticProb.addTractionToElements(elemID, trac, faceIndex,
                                            nastranOrdering=True)
-
-
-    ####### Static solver methods ########
-
-    def reset(self, SP):
-        """ Reset each of the solution to last converged value."""
-        self.setStructProblem(SP)
-        self.u.copyValues(self.u_old)
-
-    def _initializeSolve(self):
-        """
-        Initialze the solution of the structural system for the
-        loadCase. The stiffness matrix is assembled and factored.
-        """
-
-        if self._factorOnNext:
-            self.assembler.assembleJacobian(self.alpha, self.beta, self.gamma, self.res, self.K)
-            self.PC.factor()
-            self.old_update.zeroEntries()
-            self._factorOnNext = False
-            self._PCfactorOnNext = False
-
-    def __call__(self, structProblem, damp=1.0, useAitkenAcceleration=False,
-                 dampLB=0.2, loadScale=1.0):
-        """
-        Solution of the structural system for loadCase. The
-        forces must already be set.
-
-        Parameters
-        ----------
-        structProblem
-
-        Optional Arguments:
-
-        damp, float: Value to use to damp the solution update. Default is 1.0
-
-        useAitkenAcceleration, boolen: Flag to use
-        aitkenAcceleration. Only applicable for aerostructural
-        problems. Default is False.
-
-        loadScale, float: value to scale external loads by. Only useful for
-        load step approach on nonlinear problems.
-        """
-        startTime = time.time()
-
-        self.setStructProblem(structProblem)
-        self.curSP.tacsData.callCounter += 1
-
-        # Set loadScale attributes, during load incrementation, self.loadScale is the current loadScale
-        # while self.maxLoadScale is the target/final load scale.
-        # For now, maxLoadScale is set equal to self.loadScale to make _updateResidual
-        # and _getForces work, this will be addressed in future when new NL solver is merged
-        self.loadScale = loadScale
-        self.maxLoadScale = loadScale
-
-        setupProblemTime = time.time()
-
-        # Check if we need to initialize
-        self._initializeSolve()
-
-        initSolveTime = time.time()
-
-        # Compute the RHS
-        # TODO: Auxiliary forces still need to be load scaled
-        # self.structure.setLoadFactor(self.curSP.tacsData.lcnum,loadScale)
-        self.assembler.assembleRes(self.res)
-        # Zero out bc terms in F
-        self.assembler.applyBCs(self.F)
-        # Add the -F
-        self.res.axpy(-loadScale, self.F)
-
-        # Set initnorm as the norm of F
-        self.initNorm = numpy.real(self.F.norm()) * loadScale
-
-        # Starting Norm for this compuation
-        self.startNorm = numpy.real(self.res.norm())
-
-        initNormTime = time.time()
-
-        # Solve Linear System for the update
-        self.KSM.solve(self.res, self.update)
-
-        self.update.scale(-1.0)
-
-        solveTime = time.time()
-
-        # Apply Aitken Acceleration if necessary:
-        if useAitkenAcceleration:
-            if self.doDamp:
-                # Comput: temp0 = update - old_update
-                self.temp0.zeroEntries()
-                self.temp0.axpy(1.0, self.update)
-                self.temp0.axpy(-1.0, self.old_update)
-                dnom = self.temp0.dot(self.temp0)
-                damp = damp * (1.0 - self.temp0.dot(self.update) / dnom)
-
-                # Clip to a reasonable range
-                damp = numpy.clip(damp, dampLB, 1.0)
-            self.doDamp = True
-
-        # Update State Variables
-        self.assembler.getVariables(self.u)
-        self.u.axpy(damp, self.update)
-        self.assembler.setVariables(self.u)
-
-        # Set the old update
-        self.old_update.copyValues(self.update)
-
-        stateUpdateTime = time.time()
-
-        # Compute final FEA Norm
-        self.assembler.assembleRes(self.res)
-        self.res.axpy(-loadScale, self.F)  # Add the -F
-        self.finalNorm = numpy.real(self.res.norm())
-
-        finalNormTime = time.time()
-
-        # If timing was was requested print it, if the solution is nonlinear
-        # print this information automatically if prinititerations was requested.
-        if (self.getOption('printTiming') or (self.getOption('printIterations')
-                                              and self.getOption('solutionType').lower() != 'linear')):
-            self.pp('+--------------------------------------------------+')
-            self.pp('|')
-            self.pp('| TACS Solve Times:')
-            self.pp('|')
-            self.pp('| %-30s: %10.3f sec' % ('TACS Setup Time', setupProblemTime - startTime))
-            self.pp('| %-30s: %10.3f sec' % ('TACS Solve Init Time', initSolveTime - setupProblemTime))
-            self.pp('| %-30s: %10.3f sec' % ('TACS Init Norm Time', initNormTime - initSolveTime))
-            self.pp('| %-30s: %10.3f sec' % ('TACS Solve Time', solveTime - initNormTime))
-            self.pp('| %-30s: %10.3f sec' % ('TACS State Update Time', stateUpdateTime - solveTime))
-            self.pp('| %-30s: %10.3f sec' % ('TACS Final Norm Time', finalNormTime - stateUpdateTime))
-            self.pp('|')
-            self.pp('| %-30s: %10.3f sec' % ('TACS Total Solution Time', finalNormTime - startTime))
-            self.pp('+--------------------------------------------------+')
-
-        return damp
-
-    ####### Function eval/sensitivity methods ########
-
-    def evalFunctions(self, structProblem, funcs, evalFuncs=None,
-                      ignoreMissing=False):
-        """
-        This is the main routine for returning useful information from
-        pytacs. The functions corresponding to the strings in
-        EVAL_FUNCS are evaluated and updated into the provided
-        dictionary.
-
-        Parameters
-        ----------
-        structProblem : pyStructProblem class
-            Structural problem to get the solution for
-        funcs : dict
-            Dictionary into which the functions are saved.
-        evalFuncs : iterable object containing strings.
-            If not none, use these functions to evaluate.
-        ignoreMissing : bool
-            Flag to supress checking for a valid function. Please use
-            this option with caution.
-
-        Examples
-        --------
-        >>> funcs = {}
-        >>> FEAsolver(sp)
-        >>> FEAsolver.evalFunctions(sp, funcs, ['mass'])
-        >>> funcs
-        >>> # Result will look like (if structProblem, sp, has name of 'c1'):
-        >>> # {'cl_mass':12354.10}
-        """
-        startTime = time.time()
-        # Set the structural problem
-        self.setStructProblem(structProblem)
-        if evalFuncs is None:
-            evalFuncs = sorted(list(self.curSP.evalFuncs))
-        else:
-            evalFuncs = sorted(list(evalFuncs))
-
-        if not ignoreMissing:
-            for f in evalFuncs:
-                if not f in self.functionList:
-                    raise self.TACSError("Supplied function '%s' has not been added "
-                                "using addFunction()." % f)
-
-        setupProblemTime = time.time()
-
-        # Fast parallel function evaluation of structural funcs:
-        handles = [self.functionList[f] for f in evalFuncs if
-                   f in self.functionList]
-        funcVals = self.assembler.evalFunctions(handles)
-
-        functionEvalTime = time.time()
-
-        # Assign function values to appropriate dictionary
-        i = 0
-        for f in evalFuncs:
-            if f in self.functionList:
-                key = self.curSP.name + '_%s' % f
-                self.curSP.funcNames[f] = key
-                funcs[key] = funcVals[i]
-                i += 1
-
-        dictAssignTime = time.time()
-
-        if self.getOption('printTiming'):
-            self.pp('+--------------------------------------------------+')
-            self.pp('|')
-            self.pp('| TACS Function Times:')
-            self.pp('|')
-            self.pp('| %-30s: %10.3f sec' % ('TACS Function Setup Time', setupProblemTime - startTime))
-            self.pp('| %-30s: %10.3f sec' % ('TACS Function Eval Time', functionEvalTime - setupProblemTime))
-            self.pp('| %-30s: %10.3f sec' % ('TACS Dict Time', dictAssignTime - functionEvalTime))
-            self.pp('|')
-            self.pp('| %-30s: %10.3f sec' % ('TACS Function Time', dictAssignTime - startTime))
-            self.pp('+--------------------------------------------------+')
-
-    def evalFunctionsSens(self, structProblem, funcsSens, evalFuncs=None):
-        """
-        This is the main routine for returning useful (sensitivity)
-        information from pytacs. The derivatives of the functions
-        corresponding to the strings in EVAL_FUNCS are evaluated and
-        updated into the provided dictionary.
-
-        Parameters
-        ----------
-        structProblem : pyStructProblem class
-            Structural problem to get the solution for
-        funcsSens : dict
-            Dictionary into which the derivatives are saved.
-        evalFuncs : iterable object containing strings
-            The functions the user wants returned
-
-        Examples
-        --------
-        >>> funcsSens = {}
-        >>> FEAsolver.evalFunctionsSens(sp, funcsSens, ['mass'])
-        >>> funcs
-        >>> # Result will look like (if structProblem, sp, has name of 'c1'):
-        >>> # {'c1_mass':{'struct':[1.234, ..., 7.89]}
-        """
-
-        startTime = time.time()
-
-        # Set the structural problem
-        self.setStructProblem(structProblem)
-        if evalFuncs is None:
-            evalFuncs = sorted(list(self.curSP.evalFuncs))
-        else:
-            evalFuncs = sorted(list(evalFuncs))
-        # Check that the functions are all ok.
-        # and prepare tacs vecs for adjoint procedure
-        dvSenses = []
-        xptSenses = []
-        dIdus = []
-        adjoints = []
-        for f in evalFuncs:
-            if f not in self.functionList:
-                raise self.TACSError("Supplied function has not beed added "
-                            "using addFunction()")
-            else:
-                # Populate the lists with the tacs bvecs
-                # we'll need for each adjoint/sens calculation
-                dvSens = self.dvSensList[f]
-                dvSens.zeroEntries()
-                dvSenses.append(dvSens)
-
-                xptSens = self.xptSensList[f]
-                xptSens.zeroEntries()
-                xptSenses.append(xptSens)
-
-                dIdu = self.dIduList[f]
-                dIdu.zeroEntries()
-                dIdus.append(dIdu)
-
-                adjoint = self.adjointList[f]
-                adjoint.zeroEntries()
-                adjoints.append(adjoint)
-
-        setupProblemTime = time.time()
-
-        adjointStartTime = {}
-        adjointEndTime = {}
-
-        # Next we will solve all the adjoints
-        # Set adjoint rhs
-        self.addSVSens(evalFuncs, dIdus)
-        adjointRHSTime = time.time()
-        for i, f in enumerate(evalFuncs):
-            adjointStartTime[f] = time.time()
-            self.solveAdjoint(dIdus[i], adjoints[i])
-            adjointEndTime[f] = time.time()
-
-        adjointFinishedTime = time.time()
-        # Evaluate all the adoint res prooduct at the same time for
-        # efficiency:
-        self.addDVSens(evalFuncs, dvSenses)
-        self.addAdjointResProducts(adjoints, dvSenses)
-        self.addXptSens(evalFuncs, xptSenses)
-        self.addAdjointResXptSensProducts(adjoints, xptSenses)
-
-        # Recast sensititivities into dict for user
-        for i, f in enumerate(evalFuncs):
-            key = self.curSP.name + '_%s' % f
-            # Finalize sensitivity arrays across all procs
-            dvSenses[i].beginSetValues()
-            dvSenses[i].endSetValues()
-            xptSenses[i].beginSetValues()
-            xptSenses[i].endSetValues()
-            # Return sensitivities as array in sens dict
-            funcsSens[key] = {self.varName: dvSenses[i].getArray().copy(),
-                              self.coordName: xptSenses[i].getArray().copy()}
-
-        totalSensitivityTime = time.time()
-
-        if self.getOption('printTiming'):
-            self.pp('+--------------------------------------------------+')
-            self.pp('|')
-            self.pp('| TACS Adjoint Times:')
-            print('|')
-            print('| %-30s: %10.3f sec' % ('TACS Sens Setup Problem Time', setupProblemTime - startTime))
-            print('| %-30s: %10.3f sec' % (
-                'TACS Adjoint RHS Time', adjointRHSTime - setupProblemTime))
-            for f in evalFuncs:
-                print('| %-30s: %10.3f sec' % (
-                'TACS Adjoint Solve Time - %s' % (f), adjointEndTime[f] - adjointStartTime[f]))
-            print('| %-30s: %10.3f sec' % ('Total Sensitivity Time', totalSensitivityTime - adjointFinishedTime))
-            print('|')
-            print('| %-30s: %10.3f sec' % ('Complete Sensitivity Time', totalSensitivityTime - startTime))
-            print('+--------------------------------------------------+')
-
-    ####### Design variable methods ########
-
-    def setVarName(self, varName):
-        """
-        Set a name for the structural variables in pyOpt. Only needs
-        to be changed if more than 1 pytacs object is used in an
-        optimization
-
-        Parameters
-        ----------
-        varName : str
-            Name of the structural variable used in addVarGroup().
-            """
-        self.varName = varName
-
-    def setDesignVars(self, x):
-        """
-        Update the design variables used by tacs.
-
-        Parameters
-        ----------
-        x : ndarray
-            The variables (typically from the optimizer) to set. It
-            looks for variable in the ``self.varName`` attribute.
-
-        """
-        # Check if the design variables are being handed in a dict
-        if isinstance(x, dict):
-            if self.varName in x:
-                self.x.getArray()[:] = x[self.varName]
-        # or array
-        elif isinstance(x, np.ndarray):
-            self.x.getArray()[:] = x
-        else:
-            raise ValueError("setDesignVars must be called with either a numpy array or dict as input.")
-
-        # Set the variables in tacs, and the constriant objects
-        self.assembler.setDesignVars(self.x)
-        self._factorOnNext = True
-
-    def getDesignVars(self):
-        """
-        get the design variables that were specified with
-        addVariablesPyOpt.
-
-        Returns
-        ----------
-        x : array
-            The current design variable vector set in tacs.
-
-        Notes
-        -----
-        This routine **can** also accept a list or vector of
-        variables. This is used internally in pytacs, but is not
-        recommended to used externally.
-        """
-
-        # Set the variables in tacs, and the constriant objects
-        # Set the variables in tacs, and the constriant objects
-        self.assembler.getDesignVars(self.x)
-        return self.x.getArray().copy()
-
-    def getNumDesignVars(self):
-        """
-        Return the number of design variables on this processor.
-        """
-        return self.x.getSize()
-
-    def getTotalNumDesignVars(self):
-        """
-        Return the number of design variables across all processors.
-        """
-        return self.dvNum
-
-    def getCoordinates(self):
-        """
-        Return the mesh coordiantes of the structure.
-
-        Returns
-        -------
-        coords : array
-            Structural coordinate in array of size (N, 3) where N is
-            the number of structural nodes on this processor.
-        """
-        Xpts = self.assembler.createNodeVec()
-        self.assembler.getNodes(Xpts)
-        coords = Xpts.getArray()
-
-        return coords
-
-    def setCoordinates(self, coords):
-        """
-        Set the mesh coordinates of the structure.
-
-        Returns
-        -------
-        coords : array
-            Structural coordinate in array of size (N, 3) where N is
-            the number of structural nodes on this processor.
-        """
-        XptsArray = self.Xpts.getArray()
-        # Make sure input is raveled (1D) in case user changed shape
-        XptsArray[:] = numpy.ravel(coords)
-        self.assembler.setNodes(self.Xpts)
-        self._factorOnNext = True
-
-    def getNumCoordinates(self):
-        """
-        Return the number of mesh coordinates on this processor.
-        """
-        return self.Xpts.getSize()
-
-    ####### Post processing methods ########
-
-    def getVariablesAtPoints(self, structProblem, points):
-        '''The function is used to get the state variables DOF's at the
-        selected physical locations, points. A closest point search is
-        used to determine the FE nodes that are the closest to the
-        requested nodes.
-
-        NOTE: The number and units of the entries of the state vector
-        depends on the physics problem being solved and the dofs included
-        in the model.
-
-        A couple of examples of state vector components for common problem are listed below:
-
-        In Elasticity with varsPerNode = 3,
-        q = [u, v, w] # displacements
-        In Elasticity with varsPerNode = 6,
-        q = [u, v, w, tx, ty, tz] # displacements + rotations
-        In Thermoelasticity with varsPerNode = 4,
-        q = [u, v, w, T] # displacements + temperature
-        In Thermoelasticity with varsPerNode = 7,
-        q = [u, v, w, tx, ty, tz, T] # displacements + rotations + temperature
-        '''
-        try:
-            from scipy.spatial import cKDTree
-        except:
-            raise self.TACSError("scipy.spatial "
-                        "must be available to use getDisplacements")
-
-        points = numpy.atleast_2d(points)
-        self.setStructProblem(structProblem)
-
-        # Pull out the local nodes on the proc and search "points" in the tree
-        vpn = self.varsPerNode
-        Xpts = self.assembler.createNodeVec()
-        self.assembler.getNodes(Xpts)
-        localNodes = np.real(Xpts.getArray())
-        nNodes = len(localNodes) // vpn
-        xNodes = localNodes[0:nNodes * 3].reshape((nNodes, 3)).copy()
-        tree = cKDTree(xNodes)
-
-        d, index = tree.query(points, k=1)
-
-        # Now figure out which proc has the best distance for this
-        localu = np.real(structProblem.tacsData.u.getArray())
-        uNodes = localu[0:nNodes * vpn].reshape((nNodes, vpn)).copy()
-        u_req = numpy.zeros([len(points), vpn])
-        for i in range(len(points)):
-            proc = self.comm.allreduce((d[i], self.comm.rank), op=MPI.MINLOC)
-            u_req[i, :] = uNodes[index[i], :]
-            u_req[i, :] = self.comm.bcast(uNodes[index[i], :], root=proc[1])
-        return u_req
-
-    def writeDVVisualization(self, fileName, n=17):
-        """
-        This function writes a standard f5 output file, but with
-        design variables defined by x=mod(arange(nDV, n)), where n an
-        integer supplied by the user. The idea is to use contouring in
-        a post processing program to visualize the structural design
-        variables.
-
-        Parameters
-        ----------
-        fileName : str
-            Filename to use. Since it is a f5 file, shoud have .f5 extension
-        n : int
-            Modulus value. 17 is the default which tends to work well.
-            """
-
-        nDVs = self.getNumDesignVars()
-
-        # Save the current variables
-        xSave =  self.getDesignVars()
-
-        # Generate and set the 'mod' variables
-        x = numpy.mod(numpy.arange(nDVs), n)
-        self.setDesignVars(x)
-
-        # Normal solution write
-        self.writeOutputFile(fileName)
-
-        # Reset the saved variables
-        self.setDesignVars(xSave)
-
-    def writeOutputFile(self, fileName):
-        """Low-level command to write the current loadcase to a file
-
-        Parameters
-        ----------
-        fileName : str
-            Filename for output. Should have .f5 extension.
-         """
-        self.outputViewer.writeToFile(fileName)
-
-    def writeSolution(self, outputDir=None, baseName=None, number=None):
-        """This is a generic shell function that writes the output
-        file(s).  The intent is that the user or calling program can
-        call this function and pyTACS writes all the files that the
-        user has defined. It is recommneded that this function is used
-        along with the associated logical flags in the options to
-        determine the desired writing procedure
-
-        Parameters
-        ----------
-
-        outputDir : str or None
-            Use the supplied output directory
-        baseName : str or None
-            Use this supplied string for the base filename. Typically
-            only used from an external solver.
-        number : int or None
-            Use the user spplied number to index solution. Again, only
-            typically used from an external solver
-        """
-        # Check input
-        if outputDir is None:
-            outputDir = self.getOption('outputDir')
-
-        if baseName is None:
-            baseName = self.curSP.name
-
-        # If we are numbering solution, it saving the sequence of
-        # calls, add the call number
-        if number is not None:
-            # We need number based on the provided number:
-            baseName = baseName + '_%3.3d' % number
-        else:
-            # if number is none, i.e. standalone, but we need to
-            # number solutions, use internal counter
-            if self.getOption('numberSolutions'):
-                baseName = baseName + '_%3.3d' % self.curSP.tacsData.callCounter
-
-        # Unless the writeSolution option is off write actual file:
-        if self.getOption('writeSolution'):
-            base = os.path.join(outputDir, baseName) + '.f5'
-            self.outputViewer.writeToFile(base)
-
-        if self.getOption('writeBDF'):
-            base = os.path.join(outputDir, baseName) + '.bdf'
-            self.writeBDF(base)
 
     # =========================================================================
     #   The remainder of the routines should not be needed by a user
@@ -2075,162 +1366,6 @@ class pyTACS(BaseUI):
         Return number of components (property) groups found in bdf.
         """
         return self.nComp
-
-    def solveAdjoint(self, rhs, phi, damp=1.0):
-        """
-        Solve the structural adjoint.
-
-        Parameters
-        ----------
-        rhs : TACS BVec
-            right hand side vector for adjoint solve
-        phi : TACS BVec
-            BVec into which the adjoint is saved
-        damp : float
-            A damping variable for adjoint update. Typically only used
-            in multidisciplinary analysis
-        """
-
-        # First compute the residual
-        self.K.mult(phi, self.res)
-        self.res.axpy(-1.0, rhs)  # Add the -RHS
-
-        # Starting Norm for this compuation
-        self.startNorm = numpy.real(self.res.norm())
-
-        # Solve Linear System
-        zeroGuess = 0
-        self.update.zeroEntries()
-        self.KSM.solve(self.res, self.update, zeroGuess)
-
-        # Update the adjoint vector with the (damped) update
-        phi.axpy(-damp, self.update)
-
-        # Compute actual final FEA Norm
-        self.K.mult(phi, self.res)
-        self.res.axpy(-1.0, rhs)  # Add the -RHS
-        self.finalNorm = numpy.real(self.res.norm())
-
-    def getNumVariables(self):
-        """Return the number of degrees of freedom (states) that are
-        on this processor
-
-        Returns
-        -------
-        nstate : int
-            number of states.
-        """
-        return self.u.getSize()
-
-    def getVariables(self, structProblem, states=None):
-        """Return the current state values for the current
-        structProblem"""
-        self.setStructProblem(structProblem)
-
-        if states is None:
-            states = self.u.getArray().copy()
-        else:
-            states[:] = self.u.getArray()
-
-        return states
-
-    def setVariables(self, structProblem, states):
-        """ Set the structural states for current load case. Typically
-        only used for aerostructural analysis
-
-        Parameters
-        ----------
-        states : array
-            Values to set. Must be the size of getNumVariables()
-        """
-        self.setStructProblem(structProblem)
-        self.u.setValues(states)
-        self.assembler.setVariables(self.u)
-
-    def getVarsPerNodes(self):
-        """
-        Get the number of variables per node for the model.
-        """
-        if self.assembler is not None:
-            return self.varsPerNode
-        else:
-            raise self.TACSError("Assembler must be finalized before getVarsPerNodes can be called.")
-
-    def addSVSens(self, evalFuncs, dIduList):
-        """ Add the state variable sensitivity to the ADjoint RHS for given evalFuncs"""
-        funcHandles = [self.functionList[f] for f in evalFuncs if
-                       f in self.functionList]
-        self.assembler.addSVSens(funcHandles, dIduList, self.alpha, self.beta, self.gamma)
-
-    def addDVSens(self, evalFuncs, dvSensList, scale=1.0):
-        """ Add pratial sensitivity contribution due to design vars for evalFuncs"""
-        funcHandles = [self.functionList[f] for f in evalFuncs if
-                       f in self.functionList]
-        self.assembler.addDVSens(funcHandles, dvSensList, scale)
-
-    def addAdjointResProducts(self, adjointlist, dvSensList, scale=-1.0):
-        """ Add the adjoint product contribution to the design variable sensitivity arrays"""
-        self.assembler.addAdjointResProducts(adjointlist, dvSensList, scale)
-
-    def addXptSens(self, evalFuncs, xptSensList, scale=1.0):
-        """ Add pratial sensitivity contribution due to nodal coordinates for evalFuncs"""
-        funcHandles = [self.functionList[f] for f in evalFuncs if
-                       f in self.functionList]
-        self.assembler.addXptSens(funcHandles, xptSensList, scale)
-
-    def addAdjointResXptSensProducts(self, adjointlist, xptSensList, scale=-1.0):
-        """ Add the adjoint product contribution to the nodal coordinates sensitivity arrays"""
-        self.assembler.addAdjointResXptSensProducts(adjointlist, xptSensList, scale)
-
-    def getResidual(self, structProblem, res=None, Fext=None):
-        """
-        This routine is used to evaluate directly the structural
-        residual. Only typically used with aerostructural analysis.
-
-        Parameters
-        ----------
-        structProblem : pyStructProblem class
-            Structural problem to use
-        res : numpy array
-            If res is not None, place the residuals into this array.
-
-        Returns
-        -------
-        res : array
-            The same array if res was provided, (otherwise a new
-            array) with evaluated residuals
-        """
-        self.setStructProblem(structProblem)
-        self.assembler.assembleRes(self.res)
-        self.res.axpy(1.0, self.curSP.tacsData.F)  # Add the -F
-
-        if Fext is not None:
-            resArray = self.res.getArray()
-            resArray[:] -= Fext[:]
-
-        if res is None:
-            res = self.res.getArray().copy()
-        else:
-            res[:] = self.res.getArray()
-
-        return res
-
-    def getResNorms(self):
-        """Return the initial, starting and final Res Norms. Note that
-        the same norms are used for both solution and adjoint
-        computations"""
-
-        return (numpy.real(self.initNorm), numpy.real(self.startNorm),
-                numpy.real(self.finalNorm))
-
-    def zeroVectors(self):
-        """Zero all the tacs b-vecs"""
-        if not self._variablesCreated:
-            self._createVariables()
-        self.res.zeroEntries()
-        self.u.zeroEntries()
-        self.assembler.setVariables(self.u)
-        self.update.zeroEntries()
 
     def _createOutputGroups(self):
         """Automatically determine how to split out the output file
@@ -2479,97 +1614,3 @@ class pyTACS(BaseUI):
                     raise self.TACSError("Model references elements with differing numbers of variables per node (%d and %d). "
                                 "All elements must use same number of variables to be compatible."%(self.varsPerNode,
                                                                                                     elemVarsPerNode))
-
-    def _createVariables(self):
-        """Internal to create the variable required by TACS"""
-
-        if not self._variablesCreated:
-
-            self.x = self.assembler.createDesignVec()
-            self.Xpts = self.assembler.createNodeVec()
-
-            # Generic residual vector
-            self.res = self.assembler.createVec()
-
-            self.u_old = self.assembler.createVec()
-            self.old_update = self.assembler.createVec()
-            self.temp0 = self.assembler.createVec()
-
-            # Current adjoint vector
-            self.phi = self.assembler.createVec()
-
-            # Current adjoint RHS
-            self.adjRHS = self.assembler.createVec()
-
-            # Current derivative of objective wrt states
-            self.dIdu = self.assembler.createVec()
-
-            opt = self.getOption
-
-            # Tangent Stiffness --- process the ordering option here:
-            tmp = opt('orderingType').lower()
-            if tmp == 'natural':
-                ordering = tacs.TACS.NATURAL_ORDER
-            elif tmp == 'nd':
-                ordering = tacs.TACS.ND_ORDER
-            elif tmp == 'rcm':
-                ordering = tacs.TACS.RCM_ORDER
-            elif tmp == 'tacs_amd':
-                ordering = tacs.TACS.TACS_AMD_ORDER
-            elif tmp == 'multicolor':
-                ordering = tacs.TACS.MULTICOLOR_ORDER
-            else:
-                raise self.TACSError("Unrecognized 'orderingType' option value: "
-                            "'%s'. Valid values are: 'natural', 'nd', 'rcm', "
-                            "'tacs_amd', or 'multicolor'." % tmp)
-
-            self.K = self.assembler.createSchurMat(ordering)
-
-            # Additional Vecs for updates
-            self.update = self.assembler.createVec()
-
-            # Setup PCScMat and KSM solver
-            self.alpha = 1.0
-            self.beta = 0.0
-            self.gamma = 0.0
-            self.assembler.assembleJacobian(self.alpha, self.beta, self.gamma, self.res, self.K)
-
-            reorderSchur = 1
-            self.PC = tacs.TACS.Pc(self.K, lev_fill=opt('PCFillLevel'),
-                                   ratio_fill=opt('PCFillRatio'), reorder=reorderSchur)
-
-            # Operator, fill level, fill ratio, msub, rtol, ataol
-            if opt('KSMSolver').upper() == 'GMRES':
-                self.KSM = tacs.TACS.KSM(
-                    self.K, self.PC, opt('subSpaceSize'), opt('nRestarts'),
-                    opt('flexible'))
-            # TODO: Fix this
-            # elif opt('KSMSolver').upper() == 'GCROT':
-            #    self.KSM = tacs.TACS.GCROT(
-            #        self.K, self.PC, opt('subSpaceSize'), opt('subSpaceSize'),
-            #        opt('nRestarts'), opt('flexible'))
-            else:
-                raise self.TACSError("Unknown KSMSolver option. Valid options are "
-                            "'GMRES' or 'GCROT'")
-
-            self.KSM.setTolerances(self.getOption('L2ConvergenceRel'),
-                                   self.getOption('L2Convergence'))
-
-            if opt('useMonitor'):
-                self.KSM.setMonitor(tacs.TACS.KSMPrintStdout(
-                    opt('KSMSolver'), self.comm.rank, opt('monitorFrequency')))
-
-            self._variablesCreated = True
-
-
-class TACSLoadCase(object):
-    """
-    A container class for storing data related to a particular load case
-    """
-
-    def __init__(self):
-        self.F = None
-        self.u = None
-        self.auxElems = None
-        self.adjoints = {}
-        self.callCounter = -1
